refactor(ui): finalize diff row facade
