+++ src/opentui/HunkDiffBody.tsx
@@ -3,7 +3,7 @@ import { DEFAULT_TAB_WIDTH } from "../core/run/tabWidth";
 import { findMaxLineNumber } from "../ui/diff/codeColumns";
 import { buildSplitRows, buildStackRows } from "../ui/diff/diffRows";
 import { diffMessage, fitText } from "../ui/diff/plannedRowText";
-import { DiffRowView } from "../ui/diff/renderRows";
+import { DiffRowView } from "../ui/diff/DiffRowView";
 import { useHighlightedDiff } from "../ui/diff/useHighlightedDiff";
 import { resolveTheme } from "../ui/themes";
 import { toInternalDiffFile } from "./model";
